Refractored .html files into a pieced build system
Extract out:
- Common includes, navbar, etc.
- Enable straight forward editing

Then:
- Make all projects consistent with views folder, urls, etc.
- Extracted pages/general URLs (Home & Team) to there own app
- Make everything ready for editor additions
