+++ www/pages/views/__init__.py
@@ -1,0 +1,2 @@
+from pages.views.home import GetHomePage
+from pages.views.team import GetTeamPage
