+++ www/pages/views/home.py
@@ -1,0 +1,10 @@
+from django.http import HttpResponse
+from django.shortcuts import render
+from django.template import loader
+
+
+def GetHomePage(request):
+    template = loader.get_template('pages/home.html')
+    context = {
+    }
+    return HttpResponse(template.render(context, request))
