+++ www/pdfupload/urls.py
@@ -1,9 +1,7 @@
 from django.conf.urls import url
-from . import views
+from pdfupload.views import index
 
 # specify regexs for url matching
 urlpatterns = [
-    url(r'^home/$', views.home, name='home'),
-    url(r'^team/$', views.team, name='team'),
-    url(r'^$', views.index, name='default'),
+    url(r'^$', index, name='default'),
 ]
