+++ www/pdfupload/views/__init__.py
@@ -1,0 +1,1 @@
+from pdfupload.views.index import index
