+++ www/templates/layouts/base.html
@@ -1,0 +1,9 @@
+{% include "layouts/includes.html" %}
+
+<title>Watson TA</title>
+
+{% include "layouts/navbar.html" %}
+
+{% block content %}
+    {# Our sweet content goes here! #}
+{% endblock %}
