+++ www/templates/layouts/navbar.html
@@ -1,0 +1,22 @@
+<nav class="navbar navbar-toggleable-md navbar-light bg-faded">
+    <button class="navbar-toggler navbar-toggler-right" type="button" data-toggle="collapse" data-target="#navbarSupportedContent"
+        aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
+        <span class="navbar-toggler-icon"></span>
+    </button>
+
+    <a class="navbar-brand" href="/">💡 Watson-TA</a>
+
+    <div class="collapse navbar-collapse" id="navbarSupportedContent">
+        <ul class="navbar-nav mr-auto">
+            <li class="nav-item">
+                <a class="nav-link" href="/editor/">Editor</a>
+            </li>
+            <li class="nav-item">
+                <a class="nav-link" href="/pdfupload/">PDF Upload</a>
+            </li>
+            <li class="nav-item">
+                <a class="nav-link" href="/team/">The Team</a>
+            </li>
+        </ul>
+    </div>
+</nav>
