+++ www/templates/pages/home.html
@@ -1,0 +1,51 @@
+{% extends "layouts/base.html" %}
+
+{% block content %}
+<style>
+  ul {
+    list-style-type: none;
+  }
+  
+  .imageContainer {
+    background-image: url('http://i.huffpost.com/gen/2159432/images/o-EMPTY-CLASSROOM-facebook.jpg');
+  }
+  
+  .classroom img {
+    margin-top: -25px;
+    object-fit: cover;
+    width: 100%;
+    height: 300px;
+  }
+  
+  .white {
+    color: #eee;
+  }
+  
+  mark:hover {
+    background-color: #fdefa6;
+  }
+</style>
+
+
+<br>
+<div class="classroom"><img src="http://i.huffpost.com/gen/2159432/images/o-EMPTY-CLASSROOM-facebook.jpg" /></div>
+<br>
+<div class="container">
+  <div class="page-header">
+    <h1 class="">
+      <mark>Welcome to Watson-TA!</mark>
+    </h1>
+  </div>
+  <br>
+  <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.
+    Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute
+    irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat
+    cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum</p>
+  <h3>What Makes Watson-TA Unique</h3>
+  <p>dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
+    Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint
+    occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum</p>
+  <br>
+  <br>
+</div>
+{% endblock %}
