--- www/templates/pdfupload/home.html
@@ -1,87 +1,0 @@
-{# Load the tag library #}
-{% load bootstrap3 %}
-
-{# Load CSS and JavaScript #}
-<!-- {% bootstrap_css %} -->
-{% bootstrap_javascript %}
-{% load static %}
-
-<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
-<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0-alpha.6/css/bootstrap.min.css" integrity="sha384-rwoIResjU2yc3z8GV/NPeZWAv56rSmLldC3R/AZzGRnGxQQKnKkoFVhFQhNUwEyJ" crossorigin="anonymous">
-<script src="https://code.jquery.com/jquery-3.1.1.slim.min.js" integrity="sha384-A7FZj7v+d/sdmMqp/nOQwliLvUsJfDHW+k9Omg/a/EheAdgtzNs3hpfag6Ed950n" crossorigin="anonymous"></script>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/tether/1.4.0/js/tether.min.js" integrity="sha384-DztdAPBWPRXSA/3eYEEUWrWCy7G5KFbe8fFjk5JAIxUYHKkDx6Qin1DkWx51bBrb" crossorigin="anonymous"></script>
-<script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0-alpha.6/js/bootstrap.min.js" integrity="sha384-vBWWzlZJ8ea9aCX4pEW3rVHjgjt7zpkNpZk+02D9phzyeVkE+jo0ieGizqPLForn" crossorigin="anonymous"></script>
-
-
-<style>
-ul {
-  list-style-type: none;
-}
-
-.imageContainer {
-  background-image: url('http://i.huffpost.com/gen/2159432/images/o-EMPTY-CLASSROOM-facebook.jpg');
-}
-
-.classroom img {
-  margin-top:-25px;
-  object-fit: cover;
-  width:100%;
-  height:300px;
-}
-
-.white {
-  color: #eee;
-}
-
-mark:hover {
-  background-color: #fdefa6;
-}
-</style>
-
-{# Display django.contrib.messages as Bootstrap alerts #}
-{% bootstrap_messages %}
-
-{# Now load React via webpack #}
-{% load render_bundle from webpack_loader %}
-
-<title>PDF</title>
-<script src="https://fb.me/react-15.0.0.js"></script>
-<script src="https://fb.me/react-dom-15.0.0.js"></script>
-<body>
-    <nav class="navbar navbar-toggleable-md navbar-light bg-faded navbar-fixed-top">
-      <button class="navbar-toggler navbar-toggler-right" type="button" data-toggle="collapse" data-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
-        <span class="navbar-toggler-icon"></span>
-      </button>
-      <a class="navbar-brand" href="#">💡 watson-ta</a>
-      <div class="collapse navbar-collapse" id="navbarNav">
-        <ul class="navbar-nav">
-          <li class="nav-item active">
-            <a class="nav-link" href="#">home <span class="sr-only">(current)</span></a>
-          </li>
-          <li class="nav-item">
-            <a class="nav-link" href="#">nlc</a>
-          </li>
-          <li class="nav-item">
-            <a class="nav-link" href="http://localhost:8000/pdfupload/">pdf upload</a>
-          </li>
-          <li class="nav-item">
-            <a class="nav-link" href="http://localhost:8000/team/">the team</a>
-          </li>
-        </ul>
-      </div>
-    </nav>
-    <br>
-    <div class="classroom"><img src="http://i.huffpost.com/gen/2159432/images/o-EMPTY-CLASSROOM-facebook.jpg" /></div>
-    <br>
-    <div class="container">
-      <div class="page-header">
-        <h1 class=""><mark>Welcome to Watson-TA!</mark></h1>
-      </div>
-      <br>
-      <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum</p>
-      <h3>What Makes Watson-TA Unique</h3>
-      <p>dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum</p>
-    <br>
-    <br>
-  </div>
-</body>
