+++ www/templates/pdfupload/index.html
@@ -1,77 +1,35 @@
-{# Load the tag library #}
-{% load bootstrap3 %}
+{% extends "layouts/base.html" %}
 
-{# Load CSS and JavaScript #}
-{% bootstrap_javascript %}
-{% load static %}
-
-<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
-<link rel="newest stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0-alpha.6/css/bootstrap.min.css">
-<script src="https://code.jquery.com/jquery-3.1.1.slim.min.js" integrity="sha384-A7FZj7v+d/sdmMqp/nOQwliLvUsJfDHW+k9Omg/a/EheAdgtzNs3hpfag6Ed950n" crossorigin="anonymous"></script>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/tether/1.4.0/js/tether.min.js" integrity="sha384-DztdAPBWPRXSA/3eYEEUWrWCy7G5KFbe8fFjk5JAIxUYHKkDx6Qin1DkWx51bBrb" crossorigin="anonymous"></script>
-<script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0-alpha.6/js/bootstrap.min.js" integrity="sha384-vBWWzlZJ8ea9aCX4pEW3rVHjgjt7zpkNpZk+02D9phzyeVkE+jo0ieGizqPLForn" crossorigin="anonymous"></script>
-
-<link rel="stylesheet" href="{% static "pdfupload/style.css" %}">
-
-{# Display django.contrib.messages as Bootstrap alerts #}
-{% bootstrap_messages %}
-
-{# Now load React via webpack #}
+{# Load our JavaScript & React code with Webpack #}
 {% load render_bundle from webpack_loader %}
 
-<title>PDF</title>
-<script src="https://fb.me/react-15.0.0.js"></script>
-<script src="https://fb.me/react-dom-15.0.0.js"></script>
-<body>
-
-    <nav class="navbar navbar-toggleable-md navbar-light bg-faded">
-      <button class="navbar-toggler navbar-toggler-right" type="button" data-toggle="collapse" data-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
-        <span class="navbar-toggler-icon"></span>
-      </button>
-      <a class="navbar-brand" href="#">💡 watson-ta</a>
-      <div class="collapse navbar-collapse" id="navbarNav">
-        <ul class="navbar-nav">
-          <li class="nav-item">
-            <a class="nav-link" href="http://localhost:8000/home/">home <span class="sr-only">(current)</span></a>
-          </li>
-          <li class="nav-item">
-            <a class="nav-link" href="#">nlc</a>
-          </li>
-          <li class="nav-item active">
-            <a class="nav-link" href="#">pdf upload</a>
-          </li>
-          <li class="nav-item">
-            <a class="nav-link" href="http://localhost:8000/team/">the team</a>
-          </li>
-        </ul>
-      </div>
-    </nav>
-  </nav>
-
-  <br>
-  <div class="container">
+{% block content %}
+<br>
+<div class="container">
     <div class="page-header">
-      <h1>PDF Uploader</h1>
+        <h1>PDF Uploader</h1>
     </div>
 
     <section>
-      <h3>Natural Language Classifier</h3>
-      <p>Here are some static sentences that have been classified by the current version of our <strong>NLC</strong>.</p>
-      <ol class="dl-horizontal nlc-spaced">
-        {% for c in classifications %}
-            <li></span><code>{{ c.tag }}</code>
-              {% for w in c.words %}
-                {% if w.tag %}
-                  <span class={{ w.tag }}>{{ w.fragment }}</span>
-                {% else %}
-                  {{ w.fragment }}
-                {% endif %}
-              {% endfor %}
+        <h3>Natural Language Classifier</h3>
+        <p>Here are some static sentences that have been classified by the current version of our <strong>NLC</strong>.</p>
+        <ol class="dl-horizontal nlc-spaced">
+            {% for c in classifications %}
+            <li>
+                </span><code>{{ c.tag }}</code>
+                
+                {% for w in c.words %}
+                    {% if w.tag %}
+                        <span class={{ w.tag }}>{{ w.fragment }}</span>
+                    {% else %}
+                        {{ w.fragment }}
+                    {% endif %}
+                {% endfor %}
             </li>
-        {% endfor %}
-      </ol>
-      <!-- keeping for refrence, when i need to insert django template vars again (bill) -->
-      <!-- <ol class="dl-horizontal nlc-spaced">
+            {% endfor %}
+        </ol>
+        <!-- keeping for refrence, when i need to insert django template vars again (bill) -->
+        <!-- <ol class="dl-horizontal nlc-spaced">
       	{% for c in classifications %}
             <li><a class="glyphicon" data-toggle="collapse" href="#collapse{{ forloop.counter }}">&#x2b;</a></span><code>{{ c.tag }}</code>
               {% for w in c.words %}
@@ -119,21 +77,22 @@
             </div>
         {% endfor %}
       </ol> -->
-      <h3>React Component</h3>
-      <p>Below is a component to demonstrate React capability, as well as test our classifier's capability at discriminating sentences. <strong>New!</strong> Now we can now use alchemy backend to extract information from text.</p>
-      <div id="react-comp">
-        {% render_bundle 'main' %}
-      </div>
-      <div id="react-pdf">
-        {% render_bundle 'main' %}
-      </div>
-      <br>
-      <br>
+        <h3>React Component</h3>
+        <p>Below is a component to demonstrate React capability, as well as test our classifier's capability at discriminating
+            sentences. <strong>New!</strong> Now we can now use alchemy backend to extract information from text.</p>
+        <div id="react-comp">
+            {% render_bundle 'main' %}
+        </div>
+        <div id="react-pdf">
+            {% render_bundle 'main' %}
+        </div>
+        <br>
+        <br>
     </section>
-  </div>
-  <script>
+</div>
+<script>
     $(function () {
-      $('[data-toggle="tooltip"]').tooltip()
+        $('[data-toggle="tooltip"]').tooltip()
     })
-  </script>
-</body>
+</script>
+{% endblock %}
