--- www/templates/pdfupload/team.html
@@ -1,94 +1,0 @@
-{# Load the tag library #}
-{% load bootstrap3 %}
-
-{# Load CSS and JavaScript #}
-<!-- {% bootstrap_css %} -->
-{% bootstrap_javascript %}
-{% load static %}
-
-<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
-<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0-alpha.6/css/bootstrap.min.css" integrity="sha384-rwoIResjU2yc3z8GV/NPeZWAv56rSmLldC3R/AZzGRnGxQQKnKkoFVhFQhNUwEyJ" crossorigin="anonymous">
-<script src="https://code.jquery.com/jquery-3.1.1.slim.min.js" integrity="sha384-A7FZj7v+d/sdmMqp/nOQwliLvUsJfDHW+k9Omg/a/EheAdgtzNs3hpfag6Ed950n" crossorigin="anonymous"></script>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/tether/1.4.0/js/tether.min.js" integrity="sha384-DztdAPBWPRXSA/3eYEEUWrWCy7G5KFbe8fFjk5JAIxUYHKkDx6Qin1DkWx51bBrb" crossorigin="anonymous"></script>
-<script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0-alpha.6/js/bootstrap.min.js" integrity="sha384-vBWWzlZJ8ea9aCX4pEW3rVHjgjt7zpkNpZk+02D9phzyeVkE+jo0ieGizqPLForn" crossorigin="anonymous"></script>
-
-{# Display django.contrib.messages as Bootstrap alerts #}
-{% bootstrap_messages %}
-
-{# Now load React via webpack #}
-{% load render_bundle from webpack_loader %}
-
-<style>
-ul {
-  list-style-type: none;
-}
-
-.shrink img{
-  width: 32px;
-  height: 32px;
-}
-
-li.bottomborder{
-  border-bottom: 1px solid #ddd;
-  margin-bottom: 10px;
-  margin-top: 3px;
-}
-</style>
-
-<title>PDF</title>
-<script src="https://fb.me/react-15.0.0.js"></script>
-<script src="https://fb.me/react-dom-15.0.0.js"></script>
-<body>
-
-    <nav class="navbar navbar-toggleable-md navbar-light bg-faded">
-      <button class="navbar-toggler navbar-toggler-right" type="button" data-toggle="collapse" data-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
-        <span class="navbar-toggler-icon"></span>
-      </button>
-      <a class="navbar-brand" href="#">💡 watson-ta</a>
-      <div class="collapse navbar-collapse" id="navbarNav">
-        <ul class="navbar-nav">
-          <li class="nav-item">
-            <a class="nav-link" href="http://localhost:8000/home/">home <span class="sr-only">(current)</span></a>
-          </li>
-          <li class="nav-item">
-            <a class="nav-link" href="#">nlc</a>
-          </li>
-          <li class="nav-item">
-            <a class="nav-link" href="http://localhost:8000/pdfupload/">pdf upload</a>
-          </li>
-          <li class="nav-item active">
-            <a class="nav-link" href="#">the team</a>
-          </li>
-        </ul>
-      </div>
-    </nav>
-
-    <br>
-    <div class="container">
-      <div class="page-header">
-        <h1>🤘Sign of the Horns🤘</h1>
-      </div>
-        <blockquote class="blockquote">Talent wins games, but teamwork and intelligence wins championships
-        <footer class="blockquote-footer">Michael Jordan</footer>
-        </blockquote>
-
-      <p>Some current team information via <code>Github</code>.</p>
-      <ul class="media-list">
-        {% for d in data %}
-          <li class="media bottomborder">
-            <div class="media-left">
-              <a href="#">
-                <img class="shrink d-flex mr-3 media-object" src={{ d.avatar_url }} height="80" width="80">
-              </a>
-            </div>
-            <div class="media-body">
-              <h5 class="media-heading">{{ d.name }}</h5>
-              {{ d.username }}
-            </div>
-          </li>
-        {% endfor %}
-      </ul>
-    <br>
-    <br>
-  </div>
-</body>
