+++ www/watson_ta/urls.py
@@ -20,5 +20,5 @@ urlpatterns = [
     url(r'^api/', include('api.urls')),
     url(r'^pdfupload/', include('pdfupload.urls')),
     url(r'^admin/', admin.site.urls),
-    url(r'^', include('pdfupload.urls')),
+    url(r'^', include('pages.urls')),
 ]
